Bruisepack/Ointment code changes and bleeding values reduction

+++ UnityProject/Assets/Scripts/Health/Living/HealthStateMonitor.cs
@@ -15,7 +15,7 @@ public class HealthStateMonitor : ManagedNetworkBehaviour
 	bool isBreathingCache;
 	bool isSuffocatingCache;
 	int heartRateCache;
-	int bloodLevelCache;
+	float bloodLevelCache;
 	float oxygenDamageCache;
 	float toxinLevelCache;
 	bool isHuskCache;
